Add our own Duration, use it where required (#13)
The `Duration` class is a simple way for us to handle a duration in our system. Internally, it wraps an `int` representing the amount of minutes. The duration is guaranteed to be positive and can be *conveniently* compared, hashed and equalled (which the standard library's `Duration` is not). Additionally, it is immutable, which is offers us the extremely useful property of being Programmation-Fonctionnelle class friendly (Mr. Gallesio FTW).

Additionally, every usage of `DateTime startDate` - `DateTime endDate` has been replaced by `LocalDateTime start` - `Duration duration`.

For some awkward reasons this refactor does not trigger any test failure, which is a bit scary to be honest. Anyway. Every file that has been touched has been clean-ish-ed for the warnings and properly reformated.